+++ src/components/classifier/SwipeClassifier.js
